Implemented CRUD for appointments

+++ src/main/java/com/example/vallerydental/repository/DentistRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DentistRepository extends JpaRepository<Dentist, Long> {
+public interface DentistRepository extends JpaRepository<Dentist, Integer> {
 
 }
