+++ src/main/java/com/example/vallerydental/repository/ReceiptRepository.java
@@ -4,6 +4,7 @@ import com.example.vallerydental.model.Receipt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
+public interface ReceiptRepository extends JpaRepository<Receipt, Integer> {
 }
